add component buttonList
