+++ src/App.css
@@ -22,26 +22,3 @@ main {
   flex-direction: column;
   gap: 24px;
 }
-
-.btn-container {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  gap: 15px;
-  padding: 10px 0;
-}
-
-button {
-  border-radius: 8px;
-  border: 1px solid #cdcdcd;
-  padding: 0.6em 1.2em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-  background-color: #fff;
-  cursor: pointer;
-  transition: border-color 0.25s;
-}
-button:hover {
-  border-color: #333;
-}
